backend/pyproject.toml
[project]
name = "cobblers-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "anthropic>=0.49",
    "openai>=1.60",
    "google-genai>=1.0",
    "httpx>=0.28",
    "websockets>=14.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]
